src/main/test/java/com/aoc/y2023/ProblemDay8Test.java
package com.aoc.y2023;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay8Test implements ProblemDayTest {

    private ProblemDay8 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay8();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(20221);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(14616363770447L);
    }

}
